requirements.txt
avro-python3==1.8.2
fastavro
confluent-kafka[avro]
pandas
python-dateutil